Contains the following upstream commits:
- 3c2c85888846d405c189b6575a578ea9d8d072d8
- c7e6c1e88f74bb399233f573476d80e964165aee

+++ dcrjson/chainsvrcmds_test.go
@@ -64,7 +64,7 @@ func TestChainSvrCmds(t *testing.T) {
 			name: "createrawtransaction optional",
 			newCmd: func() (interface{}, error) {
 				return dcrjson.NewCmd("createrawtransaction", `[{"txid":"123","vout":1,"tree":0}]`,
-					`{"456":0.0123}`, 12312333333)
+					`{"456":0.0123}`, int64(12312333333))
 			},
 			staticCmd: func() interface{} {
 				txInputs := []dcrjson.TransactionInput{
